requirements.txt
tiktoken
firecrawl-py
openai
python-dotenv
pandas
openpyxl
scrapegraphai
"crawl4ai[all] @ git+https://github.com/unclecode/crawl4ai.git"